refactor(poincare): consistency of float types

+++ poincare/include/poincare/additional_results_helper.h
@@ -18,7 +18,7 @@ class AdditionalResultsHelper final {
 
   struct TrigonometryResults {
     UserExpression exactAngle;
-    float approximatedAngle;
+    double approximatedAngle;
     bool angleIsExact;
   };
   static TrigonometryResults TrigonometryAngleHelper(
